fix console error + fix reviews
